pyproject.toml
[tool.poetry]
name = "delayed_charging"
version = "0.1.0"
description = "Home Assistant custom component for delayed charging based on electricity prices"
authors = ["Mark Pfeifle"]
packages = [{ include = "custom_components" }]

[tool.poetry.dependencies]
python = ">=3.13.2, <4.0"
homeassistant = "^2025.8.3"
ipython = "^9.4.0"
pre-commit = "^4.2.0"
commitizen = "^4.8.3"
pytest-homeassistant-custom-component = "^0.13.272"
pytest-asyncio = "^1.0.0"

[tool.ruff]
line-length = 125
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
select = ["I00", "E", "F", "W"]
ignore = ["W292"]

[tool.black]
line-length = 125
target-version = ['py313']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.tox
  | \.venv
  | _build
  | build
  | dist
)/
'''

[tool.pyright]
include = ["custom_components", "tests"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/.venv",
    "**/.git",
    "**/.hg",
    "**/.tox",
    ".ruff_cache",
    "build",
    "dist",
]
pythonVersion = "3.13"
pythonPlatform = "Linux"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
asyncio_mode = "auto"
